+++ packages/sdk/src/send/send.js
@@ -1,3 +1,4 @@
+import {Buffer} from "@onflow/rlp"
 import {interaction, pipe} from "../interaction/interaction.js"
 import * as ixModule from "../interaction/interaction.js"
 import {invariant} from "../build/build-invariant.js"
@@ -24,5 +25,5 @@ export const send = async (args = [], opts = {}) => {
   opts.node = opts.node || (await config().get("accessNode.api"))
 
   if (Array.isArray(args)) args = pipe(interaction(), args)
-  return sendFn(await resolveFn(args), {config, response, ix: ixModule}, opts)
+  return sendFn(await resolveFn(args), {config, response, ix: ixModule, Buffer}, opts)
 }