+++ packages/transport-http/CHANGELOG.md
@@ -1,3 +1,8 @@
+### Unreleased
+
+- 2022-02-25 -- Use node require to consume node standard libraries
+- 2022-02-11 -- Uses Buffer as provided by context injection
+
 ### 0.0.6 -- 2022-02-04
 
 - 2022-02-04 -- Remove `0x` prefix from public keys in get account response.